Refactorizando el Repo de horas
