--- sprint 2/BackendGeems/BackendGeems/API/CalcularPago.cs
@@ -1,1 +1,0 @@
-﻿
