+++ sprint 2/BackendGeems/BackendGeems/appsettings.json
@@ -7,6 +7,6 @@
     },
     "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Data Source=ESTEBAN;Initial Catalog=GEEMSDB;Integrated Security=True;Connect Timeout=30;Encrypt=False;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False"
+    "DefaultConnection": "Data Source=DESKTOP-BLGI7EI;Initial Catalog=GEEMSDB;Integrated Security=True;Encrypt=False"
   }
 }
