+++ sprint 2/BackendGeems/UnitTests/CalculosTest.cs
@@ -6,19 +6,16 @@ namespace UnitTests;
 
 public class CalculosTest
 {
-
-    GEEMSRepo _repoInfrastructure;
     [SetUp]
     public void Setup()
     {
-        _repoInfrastructure = new GEEMSRepo();
     }
 
     [Test]
     public void CalcularImpuestoRenta_DeberiaRetornarImpuestoCorrecto()
     {
         // Arrange
-        var repo = new GEEMSRepo();
+        var repo = new GEEMSPagoRepo();
         int ingresoMensual = 1352000;
         decimal impuestoEsperado = 43000; 
 
@@ -34,7 +31,7 @@ public class CalculosTest
         public void GenerarPagoEmpleado_FechaInicioPosteriorAFechaFinal_DeberiaLanzarExcepcion()
         {
             // Arrange
-            var repo = new GEEMSRepo();
+            var repo = new GEEMSPagoRepo();
             Guid idEmpleado = Guid.NewGuid();
             Guid idPlanilla = Guid.NewGuid();
             DateTime fechaInicio = new DateTime(2024, 7, 1);
@@ -49,7 +46,7 @@ public class CalculosTest
     public void GenerarPagoEmpleado_EmpleadoInexistente_DeberiaLanzarExcepcion()
     {
         // Arrange
-        var repo = new GEEMSRepo();
+        var repo = new GEEMSPagoRepo();
         Guid idEmpleado = Guid.NewGuid();
         Guid idPlanilla = Guid.NewGuid();
         DateTime fechaInicio = new DateTime(2024, 5, 1);
